Makefile: Move -Wno-unused-but-set-variable out of GCC only block
commit 885480b084696331bea61a4f7eba10652999a9c1 upstream.

Currently, -Wunused-but-set-variable is only supported by GCC so it is
disabled unconditionally in a GCC only block (it is enabled with W=1).
clang currently has its implementation for this warning in review so
preemptively move this statement out of the GCC only block and wrap it
with cc-disable-warning so that both compilers function the same.

Cc: stable@vger.kernel.org
Link: https://reviews.llvm.org/D100581
Reviewed-by: Nick Desaulniers <ndesaulniers@google.com>
Tested-by: Nick Desaulniers <ndesaulniers@google.com>
[nc: Backport, workaround lack of e2079e93f562 in older branches]

Change-Id: I5df95411de20b0da760bcb3be48690e3f9268962
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>